+++ roles/web-app-flowise/files/playwright/test-guest-persona.js
@@ -1,0 +1,8 @@
+const { test } = require("@playwright/test");
+const { runGuestFlow } = require("./personas");
+
+test.use({ ignoreHTTPSErrors: true });
+
+test("guest: public-landing → auth chain → never authenticated", async ({ page }) => {
+  await runGuestFlow(page);
+});
