+++ roles/web-app-opentalk/files/playwright/test-biber-persona.js
@@ -1,0 +1,9 @@
+const { test } = require("@playwright/test");
+
+const { runBiberFlow } = require("./personas");
+
+test.use({ ignoreHTTPSErrors: true });
+
+test("biber: app → universal logout", async ({ page }) => {
+  await runBiberFlow(page);
+});
